+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support `Float64` (OME `double`) datatype by casting array data to `Float32`. 
+- `onHover` prop for `VivViewer`, `PictureInPictureViewer`, `SideBySideViewer` for deck.gl callback.
 
 ### Changed